+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Master
 
+- [FEATURE] All actions (onchange, onkeydown and onfocus) now receive a richer public API object
+  that is identical in shape to the one they received before but also contains `highlight(option)`
+  and `search(term)` actions
+
 # 0.7.0-beta.1
 - [BREAKING] Update to ember-basic-dropdown 0.7.0-beta.1. This means that the component is opened/
   closed using mousedown instead of click. This makes the component feel more snappy. It is unlikeliy